feat: 실행 & 테스트 관련 app 모듈 구현
