+++ kitchenpos-app/src/test/resources/h2-truncate.sql
